Unfactor NavList into separate components for the BurgerMenu and the Navigation components.
