+++ src/components/grid-aware/Navigation/BurgerMenu.tsx
@@ -1,14 +1,56 @@
+import { Link } from "gatsby";
 import * as React from "react";
 import { push as Menu } from "react-burger-menu";
 
 import s from "./BurgerMenu.module.css";
-import { NavLinkProps, NavList, ThemeType } from "./NavList";
-import navListStyles from "./NavList.BurgerMenu.module.css";
+import { NavigationItem } from "./types";
+
+/** An internal, external, or button link. */
+const NavLink = ({
+  text,
+  internalLink,
+  externalLink,
+  isButton = false,
+}: NavigationItem) => {
+  const className = `${s.navLink} ${isButton ? s.button : ""}`;
+  if (internalLink) {
+    // Enable the "active" style for any nested pages, except for the home page,
+    // which would be a parent page for any page. This is also used to get
+    // around some annoying behavior where a URL will not be considered the
+    // "current" page even if it is off by a trailing slash (e.g. "/about" !===
+    // "/about/" when it comes to computing "activeness").
+    // https://github.com/reach/router/issues/231
+    const isPartiallyActive = internalLink.replace(/\/$/, "") !== "";
+    return (
+      <Link
+        className={className}
+        activeClassName={s.currentPage}
+        to={internalLink}
+        partiallyActive={isPartiallyActive}
+      >
+        {text}
+      </Link>
+    );
+  }
+  if (externalLink) {
+    return (
+      <a
+        className={className}
+        href={externalLink}
+        target="_blank"
+        rel="noreferrer"
+      >
+        {text}
+      </a>
+    );
+  }
+  throw new Error("NavLink missing required link prop.");
+};
 
 type BurgerMenuProps = {
   pageWrapperID: string;
   outerContainerID: string;
-  navigationItems: NavLinkProps[];
+  navigationItems: NavigationItem[];
   isOpen: boolean;
   setIsOpen: (newValue: boolean) => void;
 };
@@ -36,8 +78,18 @@ const BurgerMenu = ({
     right
     width="286px"
   >
-    {/* Note: The type assertion as ThemeType is necessary because our CSS modules shim for TypeScript doesn't know about the specific keys that are accessible. */}
-    <NavList items={navigationItems} theme={navListStyles as ThemeType} />
+    <ul className={s.navList}>
+      {navigationItems.map(({ text, internalLink, externalLink, isButton }) => (
+        <li key={text} className={s.navListItem}>
+          <NavLink
+            text={text}
+            internalLink={internalLink}
+            externalLink={externalLink}
+            isButton={isButton}
+          />
+        </li>
+      ))}
+    </ul>
   </Menu>
 );
