--- src/components/grid-aware/Navigation/NavList.BurgerMenu.module.css
@@ -1,29 +1,0 @@
-.navList {
-  font: var(--font-subtitle);
-  margin: 0;
-  padding: 0;
-}
-
-.navListItem {
-  display: block;
-}
-
-.navLink {
-  color: var(--color-gray-700);
-  display: block;
-  padding: 14px 20px;
-  text-decoration: none;
-}
-
-.navLink.currentPage {
-  color: var(--color-sheltertech-blue);
-}
-
-.navLink:hover,
-.navLink.currentPage:hover,
-.navLink:active,
-.navLink.currentPage:active {
-  background-color: var(--color-sheltertech-blue);
-  color: var(--color-white);
-}
-
