+++ src/components/grid-aware/Navigation/Navigation.module.css
@@ -86,3 +86,48 @@
    */
   vertical-align: middle;
 }
+
+.navList {
+  font: var(--font-caption);
+  margin: 0;
+  padding: 0;
+}
+
+.navListItem {
+  display: inline-block;
+}
+
+.navLink {
+  color: var(--color-gray-700);
+  display: inline-block;
+  padding: 10px 20px;
+  text-decoration: none;
+}
+
+.navLink:hover {
+  background-color: var(--color-gray-300);
+  color: var(--color-black);
+}
+
+.navLink.currentPage {
+  color: var(--color-sheltertech-blue);
+}
+
+.navLink.button {
+  background-color: var(--color-sheltertech-blue);
+  color: var(--color-white);
+  margin-left: 20px;
+  padding: 10px 20px;
+  text-decoration: none;
+}
+
+.navLink.button:hover {
+  background-color: var(--color-sheltertech-blue);
+  color: var(--color-white);
+  position: relative;
+  top: 1px;
+}
+
+.navLink.button.currentPage {
+  color: var(--color-white);
+}
