+++ src/components/grid-aware/Navigation/types.ts
@@ -1,0 +1,6 @@
+export type NavigationItem = {
+  text: string;
+  internalLink?: string;
+  externalLink?: string;
+  isButton?: boolean;
+};
